Migrate demo helper functions to new demo_helper.py
