--- fizzbuzz.py
@@ -1,24 +1,0 @@
-"""
-Helper functions for fizzbuzz.ipynb.
-"""
-from typing import List
-
-def fizz_buzz_encode(x: int) -> List[int]:
-    """
-    Return a categorical encoding of a correct FizzBuzz output.
-    """
-    if x % 15 == 0:
-        return [0, 0, 0, 1]
-    elif x % 5 == 0:
-        return [0, 0, 1, 0]
-    elif x % 3 == 0:
-        return [0, 1, 0, 0]
-    else:
-        return [1, 0, 0, 0]
-
-
-def binary_encode(x: int) -> List[int]:
-    """
-    10 digit binary encoding of x
-    """
-    return [x >> i & 1 for i in range(10)]
